refactor(ett): extract user component
